Currency Converter, Reflection  & Prep
This is my first time using libraries tkinter and requests. I had a lot of fun, and learned even more. To expand this project, I'd to utilize the additional info retrieved from the API to maybe make options for visualizing data on currency rate changes.

Below is the unfiltered info on my prep for the project. I think the model behind it - ro include both a breakdown on what the program will need to run and what will happen when program runs - worked very well. Might have to work a bit more on it for projects of bigger sizes.

"""
For this project, I will need:
* A GUI with:
    3 input fields that will make the following variables:
    -from 
    -amount
    -to
    Button to retrieve the result:
    -convert

* An API to retrieve current currencies
    * A way to implement API function to GUI app
        -I will need to GET the api result 
            -it will be based on specified from, to and amount
            -the from, to and amount are specified in user input fields 
        -I can access the result with .json() method
    * A way to show the results from API in GUI app
        -I will need to print the result onto GUI 

Step-by-step:
1. User inputs from, to & amount
1.1. User clicks on convert
2. get_info(to, from, amount) Function returns the info from the API as json object with from.get(), to.get() & amount.get() variables as input
2.2. Check status code to see if API was retrieved successfully
3. get_result function returns the result key,value from get_info ##removed this step while coding, actually seemed a bit obsolete with convert function
4. convert Function shows result on GUI program
"""